+++ docs/news.rst
@@ -10,6 +10,7 @@ Lab
 * Use absolute paths in Slurm scripts to safeguard against ``.bashrc`` changing directories (Jendrik Seipp).
 * Use 9 GiB memory limit for Slurm jobs on Tetralith by default (Jendrik Seipp).
 * Revise documentation about memory limits (Jendrik Seipp).
+* Remove soft memory limit from Slurm jobs that used to guard against an old problem of the cgroup mechanism failing (Jendrik Seipp).
 
 Downward Lab
 ^^^^^^^^^^^^