+++ docs/news.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-v8.4 (unreleased)
+v8.4 (2025-02-11)
 -----------------
 
 Lab
@@ -11,9 +11,9 @@ Lab
 * Use 9 GiB memory limit for Slurm jobs on Tetralith by default (Jendrik Seipp).
 * Revise documentation about memory limits (Jendrik Seipp).
 * Remove soft memory limit from Slurm jobs that used to guard against an old problem of the cgroup mechanism failing (Jendrik Seipp).
-* Raise minimum supported Python version to 3.8 (Jendrik Seipp).
 * Remove documentation about obsolete ``infai_1`` partition in Basel (Florian Pommerening).
 * Add :meth:`SlurmEnvironment.is_present() <lab.environments.SlurmEnvironment.is_present>` method for testing whether the script is run within a given grid environment (Jendrik Seipp).
+* Raise minimum supported Python version to 3.8 (Jendrik Seipp).
 
 Downward Lab
 ^^^^^^^^^^^^